Updated flake and improved installation

+++ install.sh
@@ -2,45 +2,57 @@
 
 # Check if running as root. If root, script will exit.
 if [[ $EUID -eq 0 ]]; then
-	echo "This script should not be executed as root! Exiting..."
-	exit 1
+  echo "This script should not be executed as root! Exiting..."
+  exit 1
 fi
 
 # Check if using NixOS. If not using NixOS, script will exit.
-if [[ ! "$(grep -i nixos < /etc/os-release)" ]]; then
-	echo "This installation script only works on NixOS! Download an iso at https://nixos.org/download/"
+if [[ ! "$(grep -i nixos </etc/os-release)" ]]; then
+  echo "This installation script only works on NixOS! Download an iso at https://nixos.org/download/"
   echo "Keep in mind that this script is not intended for use while in the live environment."
-	exit 1
+  exit 1
 fi
 
 scriptdir=$(realpath "$(dirname "$0")")
 currentUser=$(logname)
 
-pushd "$scriptdir"&> /dev/null || exit
+pushd "$scriptdir" &>/dev/null || exit
 
-# Delete dirs that conflict with home-manager
-sudo rm -f ~/.mozilla/firefox/profiles.ini
-sudo rm -rf ~/.gtkrc-*
-sudo rm -rf ~/.config/gtk-*
-sudo rm -rf ~/.config/cava
+# Delete dirs that conflict with home-manager (skip symlinks)
+paths=(
+  ~/.mozilla/firefox/profiles.ini
+  ~/.gtkrc-*
+  ~/.config/gtk-*
+  ~/.config/cava
+)
+for file in "${paths[@]}"; do
+  for expanded in $file; do
+    if [ -e "$expanded" ] && [ ! -L "$expanded" ]; then
+      # echo "Removing: $expanded"
+      sudo rm -rf "$expanded"
+    fi
+  done
+done
 
 # replace username variable in flake.nix with $USER
 sed -i -e 's/username = \".*\"/username = \"'$currentUser'\"/' "$scriptdir/flake.nix"
 
 # rm -f $scriptdir/hosts/Default/hardware-configuration.nix &>/dev/null
-if [ -f "/etc/nixos/hardware-configuration.nix" ]; then
-  for host in "$scriptdir"/hosts/*/ ; do
-    host=${host%*/}
-    cat "/etc/nixos/hardware-configuration.nix" > "$host/hardware-configuration.nix"
-  done
-else
-	# Generate new config
-	clear
-  nix develop --command bash -c "echo GENERATING CONFIG! | figlet -cklno | lolcat -F 0.3 -p 2.5 -S 300"	
-  for host in "$scriptdir"/hosts/*/ ; do
-    host=${host%*/}
-    sudo nixos-generate-config --show-hardware-config > "$host/hardware-configuration.nix"
-  done
+if [ ! -f "$scriptdir/hosts/Default/hardware-configuration.nix" ]; then
+  if [ -f "/etc/nixos/hardware-configuration.nix" ]; then
+    for host in "$scriptdir"/hosts/*/; do
+      host=${host%*/}
+      cat "/etc/nixos/hardware-configuration.nix" >"$host/hardware-configuration.nix"
+    done
+  else
+    # Generate new config
+    clear
+    nix develop --command bash -c "echo GENERATING CONFIG! | figlet -cklno | lolcat -F 0.3 -p 2.5 -S 300"
+    for host in "$scriptdir"/hosts/*/; do
+      host=${host%*/}
+      sudo nixos-generate-config --show-hardware-config >"$host/hardware-configuration.nix"
+    done
+  fi
 fi
 
 nix develop --command bash -c "git -C $scriptdir add hosts/Default/hardware-configuration.nix"
@@ -51,4 +63,4 @@ nix develop --command bash -c "sudo nixos-rebuild switch --flake "$scriptdir#Def
 echo "success!"
 echo "Make sure to reboot if this is your first time using this script!"
 
-popd "$scriptdir" &> /dev/null || exit
+popd "$scriptdir" &>/dev/null || exit
